Massive refactoring: Extracted NormalAuthor and VerifiedAuthor with own ID-objects

+++ municipalityinitiative-web/src/test/java/fi/om/municipalityinitiative/service/UserServiceIntegrationTest.java
@@ -83,7 +83,7 @@ public class UserServiceIntegrationTest extends ServiceIntegrationTestBase{
         userService.authorLogin(testHelper.getPreviousTestManagementHash(), requestMock);
         LoginUserHolder loginUserHolder = userService.getRequiredLoginUserHolder(requestMock);
 
-        assertThat(loginUserHolder.getNormalLoginUser().getAuthorId(), is(testHelper.getLastAuthorId()));
+        assertThat(loginUserHolder.getNormalLoginUser().getAuthorId().toLong(), is(testHelper.getLastAuthorId()));
         assertThat(loginUserHolder.getUser().hasRightToInitiative(initiative), is(true));
         assertThat(loginUserHolder.getUser().hasRightToInitiative(-1L), is(false));
     }
